+++ src/anime_parsers_ru/parser_aniboom.py
@@ -29,11 +29,11 @@ class AniboomParser:
         if mirror: # Если есть зеркало, то меняем домен на него
             self._dmn = mirror
         else:
-            self._dmn = "animego.org"
+            self._dmn = "animego.me"
 
     def fast_search(self, title: str) -> list[dict]:
         """
-        Быстрый поиск через animego.org
+        Быстрый поиск через animego.me
         
         :title: Название аниме
 
@@ -79,7 +79,7 @@ class AniboomParser:
     
     def search(self, title: str) -> list[dict]:
         """
-        Расширенный поиск через animego.org. Собирает дополнительные данные об аниме.
+        Расширенный поиск через animego.me. Собирает дополнительные данные об аниме.
 
         :title: Название
 
@@ -141,7 +141,7 @@ class AniboomParser:
         """
         Возвращает данные по эпизодам.
 
-        :link: ссылка на страницу с данными (прим: https://animego.org/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546)
+        :link: ссылка на страницу с данными (прим: https://animego.me/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546)
 
         Возвращает отсортированный по номеру серии список словарей в виде:
         [
@@ -180,9 +180,9 @@ class AniboomParser:
 
     def anime_info(self, link: str) -> dict:
         """
-        Получение данных об аниме с его страницы на animego.org.
+        Получение данных об аниме с его страницы на animego.me.
 
-        :link: Ссылка на страницу (прим: https://animego.org/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546)
+        :link: Ссылка на страницу (прим: https://animego.me/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546)
 
         Возвращает словарь следующего вида:
         {
@@ -288,7 +288,7 @@ class AniboomParser:
         """
         Получает информацию о переводах и их id для плеера aniboom
 
-        :animego_id: id аниме на animego.org
+        :animego_id: id аниме на animego.me
 
         Возвращает список словарей:
         [
@@ -342,7 +342,7 @@ class AniboomParser:
         """
         Возвращает ссылку на embed от aniboom. Сама по себе ссылка не может быть использована, однако требуется для дальнейшего парсинга.
 
-        :animego_id: id аниме на animego.org
+        :animego_id: id аниме на animego.me
 
         Возвращает ссылку в виде: https://aniboom.one/embed/yxVdenrqNar
         Если ссылка не найдена, выкидывает NoResults exception
@@ -398,7 +398,7 @@ class AniboomParser:
     
     def _get_media_src(self, embed_link: str, episode: int, translation: str) -> str:
         """
-        Возвращает ссылку на mpd файл. Скачивать файл нужно обязательно с header'ами что отправлен запрос от animego.org или aniboom.one (иначе 403 ошибка)
+        Возвращает ссылку на mpd файл. Скачивать файл нужно обязательно с header'ами что отправлен запрос от animego.me или aniboom.one (иначе 403 ошибка)
 
         :embed_link: ссылка на embed (можно получить из _get_embed_link)
         :episode: Номер эпизода (вышедшего) (Если фильм - 0)
@@ -458,16 +458,15 @@ class AniboomParser:
         }
         playlist = requests.get(media_src, headers=headers).text
         # Вставляем полный путь до сервера
-        filename = media_src[media_src.rfind('/')+1:media_src.rfind('.')]
-        server_path = media_src[:media_src.rfind('.')]
-        playlist = playlist.replace(filename, server_path)
+        server_path = media_src[:media_src.rfind('master_device.m3u8')]
+        playlist = playlist.replace('media_', server_path+'media_')
         return playlist
     
     def get_mpd_playlist(self, animego_id: str, episode: int, translation_id: str) -> str:
         """
         Возвращает mpd файл строкой (содержимое файла)
         
-        :animego_id: id аниме на animego.org (может быть найдена из fast_search по ключу 'animego_id' для нужного аниме или из search по ключу 'animego_id' для нужного аниме) (из ссылки на страницу аниме https://animego.org/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546 > 2546)
+        :animego_id: id аниме на animego.me (может быть найдена из fast_search по ключу 'animego_id' для нужного аниме или из search по ключу 'animego_id' для нужного аниме) (из ссылки на страницу аниме https://animego.me/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546 > 2546)
         :episode: Номер эпизода (вышедшего) (Если фильм - 0)
         :translation_id: id перевода (который именно для aniboom плеера) (можно получить из get_translations_info)
 
@@ -483,7 +482,7 @@ class AniboomParser:
         """
         Сохраняет mpd файл как указанный filename
         
-        :animego_id: id аниме на animego.org (может быть найдена из fast_search по ключу 'animego_id' для нужного аниме или из search по ключу 'animego_id' для нужного аниме) (из ссылки на страницу аниме https://animego.org/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546 > 2546)
+        :animego_id: id аниме на animego.me (может быть найдена из fast_search по ключу 'animego_id' для нужного аниме или из search по ключу 'animego_id' для нужного аниме) (из ссылки на страницу аниме https://animego.me/anime/volchica-i-pryanosti-torgovec-vstrechaet-mudruyu-volchicu-2546 > 2546)
         :episode: Номер эпизода (вышедшего) (Если фильм - 0)
         :translation_id: id перевода (который именно для aniboom плеера) (можно получить из get_translations_info)
         :filename: Имя/путь для сохраняемого файла обязательно чтобы было .mpd расширение (прим: result.mpd или content/result.mpd)